Filters: introduce a module for code-specialized biquad filters
These are optimized for the case of the external consumer and crossfade between
two filters to avoid instability.
As with other parts of Synthizer we actually use templates to implement
specialized versions: this is because any multiple of 4 should be autovectorized
at least to some extent, and also because in practice the same channel widths
will tend to be used (e.g. 1, 2, 4, 6, 8 are the most common).
We can revisit this later if it proves not to help, but in the steady state
any penalty from the increased code size should be outweighed by the savings.

at some point, we will want to verify that the above supposition is correct,
but this implementation isn't very hard especially since using tempaltes lets
us reuse the already-proven IIRFilter templates.

Refs #25